+++ package.json
@@ -34,6 +34,7 @@
     "typescript": "^3.0.3"
   },
   "dependencies": {
+    "commander": "^6.1.0",
     "ethers": "^4.0.0-beta.14",
     "fs-extra": "^8.1.0",
     "web3-core-helpers": "^1.2.11",
