+++ src/cmds/index.ts
@@ -1,0 +1,7 @@
+import {evmCommand} from "./evm";
+import {rpcCommand} from "./rpc";
+
+export {
+    evmCommand,
+    rpcCommand
+}
